requirements.txt
# GPT-Researcher Project Requirements
# Note: GPT-Researcher is installed via git, but here are the key dependencies

gpt-researcher>=0.14.5
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
asyncio
python-dotenv
